Finalized the types and update components accordingly
Since the card and folder functionality are complete. I refined and added functions that I thought would be nice to have.

I decided to remove the UniqueIdentifier from the types as it's nice to know what it can actually contain. And it would probably help if I were to remove the dnd-kit for another package.

The names are still working in progress, but so far I'm satisfied and ready to work on the Binder system.