Add Temp. Drag Handle to Fix the No Space in Textarea issue
From what I've understood, this is a dnd-kit issue. I figured it out that assigning the listeners to the entire Card would cause the textarea preventing to create spaces of the program.

It's apparently interfering with the onChange event: https://stackoverflow.com/questions/77415442/listeners-from-dnd-kit-are-interfering-with-the-inputcheckboxs-onchange-event.

However, I have limited knowledge of how it works, but it's something along the lines with that.

Anyhow, I implemented the Drag Handle to not be that intrusive. I might make it a Component and put it to other Draggables.